* refactoring and splitting a DI module
