+++ GamePortal/AliaksNad.Battleship.Logic/DIModules/ContextModule.cs
@@ -1,0 +1,13 @@
+﻿using AliaksNad.Battleship.Data.Contexts;
+using Ninject.Modules;
+
+namespace AliaksNad.Battleship.Logic.DIModules
+{
+    public class ContextModule : NinjectModule
+    {
+        public override void Load()
+        {
+            this.Bind<BattleAreaContext>().ToSelf();
+        }
+    }
+}
